requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
astroid==2.3.3
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.4
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
tzlocal==2.0.0
Werkzeug==1.0.0
wrapt==1.11.2
